pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[project]
name = "chat-hub"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqliter-py>=0.3",
]

[project.optional-dependencies]
protocol = ["chat-hub-protocol"]
dev = [
    "ruff>=0.4",
]

# ── 工具配置 ──────────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B"]
